feat(auth): refactor auth page
